refactor: 패키지 이동 및 메서드 분리
